- Convert flashing and dumping functions from WorkerForm into TaskerForm tasks - Convert hmod functions into TaskerForm tasks - Install hmod immediately after transferring instead of transferring the entire queue and installing all at once - Embed certain files into exe including the base hmods - Download kernel from hakchi.net - Change SSH.NET stdin buffer size to 128KB - Update splash

+++ Tasks/TaskerForm.cs
@@ -36,7 +36,7 @@ namespace com.clusterrr.hakchi_gui.Tasks
 
         // helper combined setters
 
-        public void SetProgress(int value, int maxValue, State state, string status)
+        public void SetProgress(long value, long maxValue, State state, string status)
         {
             SetProgress(value, maxValue);
             SetState(state);
@@ -96,7 +96,7 @@ namespace com.clusterrr.hakchi_gui.Tasks
             return State.Undefined;
         }
 
-        public void SetProgress(int value, int maxValue)
+        public void SetProgress(long value, long maxValue)
         {
             if (value < 0 || maxValue < 0) return;
             if (Disposing) return;
